fix: Add projectUuid authentication to DOCX hook and all download URLs
Issue: DOCX files failing to load with 'Failed to fetch DOCX file' error
Root cause: useDocxContent hook was using wrong API endpoint without auth

Changes:
✅ **Fixed DOCX Hook Authentication**
- Updated useDocxContent to use /api/library/download/ endpoint
- Added projectUuid parameter for proper authentication
- Fixed dependency array to include projectUuid

✅ **Restored Download URL Authentication**
- Re-added currentProject.uuid to all download URLs
- Fixed PDF, image, text, and DOCX file access
- Consistent authentication across all preview types

✅ **Maintained Hook Benefits**
- Kept DOCX logic in separate hook (organization benefit)
- Added proper authentication (functionality fix)
- No change to component structure or user experience

Result: DOCX files now load correctly with proper authentication.
All file previews (PDF, images, text, DOCX) work in live environment.
